Centralizes BaSyx backend component integration in store
