+++ aas-web-ui/src/store/NavigationStore.ts
@@ -1,11 +1,24 @@
+import type { AutoSyncType, PlatformType, PluginType, SnackbarType } from '@/types/Application';
+import type { BaSyxComponent, RepositoryKey } from '@/types/BaSyx';
 import { defineStore } from 'pinia';
-import { computed, ref } from 'vue';
+import { computed, reactive, ref } from 'vue';
 import { LocationQuery } from 'vue-router';
-import { AutoSyncType, PlatformType, PluginType, SnackbarType } from '@/types/Application';
-import { useAASStore } from './AASDataStore';
+import { useRequestHandling } from '@/composables/RequestHandling';
+import { useAASStore } from '@/store/AASDataStore';
+import { useEnvStore } from '@/store/EnvironmentStore';
 
 export const useNavigationStore = defineStore('navigationStore', () => {
-    // States
+    // Initialize Dependent Stores
+    const aasStore = useAASStore();
+    const envStore = useEnvStore();
+
+    // Composables
+    const { getRequest } = useRequestHandling();
+
+    // Computed Property
+    const endpointConfigAvailable = computed(() => envStore.getEndpointConfigAvailable);
+
+    // State Variables
     const drawerState = ref(true);
     const AASDiscoveryURL = ref('');
     const AASRegistryURL = ref('');
@@ -13,15 +26,61 @@ export const useNavigationStore = defineStore('navigationStore', () => {
     const AASRepoURL = ref('');
     const SubmodelRepoURL = ref('');
     const ConceptDescriptionRepoURL = ref('');
-    const Snackbar = ref({} as SnackbarType);
-    const AutoSync = ref({} as AutoSyncType);
+    const Snackbar = ref<SnackbarType>({} as SnackbarType);
+    const AutoSync = ref<AutoSyncType>({} as AutoSyncType);
     const StatusCheck = ref(false);
     const isMobile = ref(false);
-    const platform = ref({} as PlatformType);
-    const plugins = ref([] as PluginType[]);
+    const platform = ref<PlatformType>({} as PlatformType);
+    const plugins = ref<PluginType[]>([]);
     const triggerAASListReload = ref(false);
     const triggerAASListScroll = ref(false);
-    const urlQuery = ref({} as LocationQuery);
+    const urlQuery = ref<LocationQuery>({} as LocationQuery);
+
+    // Reactive BaSyx Components Configurations
+    const basyxComponents = reactive<Record<RepositoryKey, BaSyxComponent>>({
+        AASDiscovery: {
+            url: AASDiscoveryURL,
+            loading: ref(false),
+            connected: ref(null),
+            connect: () => connectComponent('AASDiscovery'),
+            label: 'AAS Discovery URL',
+        },
+        AASRegistry: {
+            url: AASRegistryURL,
+            loading: ref(false),
+            connected: ref(null),
+            connect: () => connectComponent('AASRegistry'),
+            label: 'AAS Registry URL',
+        },
+        SubmodelRegistry: {
+            url: SubmodelRegistryURL,
+            loading: ref(false),
+            connected: ref(null),
+            connect: () => connectComponent('SubmodelRegistry'),
+            label: 'Submodel Registry URL',
+        },
+        AASRepo: {
+            url: AASRepoURL,
+            loading: ref(false),
+            connected: ref(null),
+            connect: () => connectComponent('AASRepo'),
+            label: 'AAS Repository URL',
+        },
+        SubmodelRepo: {
+            url: SubmodelRepoURL,
+            loading: ref(false),
+            connected: ref(null),
+            connect: () => connectComponent('SubmodelRepo'),
+            label: 'Submodel Repository URL',
+        },
+        ConceptDescriptionRepo: {
+            url: ConceptDescriptionRepoURL,
+            loading: ref(false),
+            connected: ref(null),
+            connect: () => connectComponent('ConceptDescriptionRepo'),
+            label: 'Concept Description Repository URL',
+        },
+    });
 
     // Getters
     const getDrawerState = computed(() => drawerState.value);
@@ -40,34 +99,9 @@ export const useNavigationStore = defineStore('navigationStore', () => {
     const getTriggerAASListReload = computed(() => triggerAASListReload.value);
     const getTriggerAASListScroll = computed(() => triggerAASListScroll.value);
     const getUrlQuery = computed(() => urlQuery.value);
+    const getBasyxComponents = computed(() => basyxComponents);
 
     // Actions
-    function dispatchComponentURL(componentKey: string, url: string, clearSelectedNode: boolean = true): void {
-        // console.log('dispatchComponentURL', componentKey, url, clearSelectedNode);
-        switch (componentKey) {
-            case 'AASDiscovery':
-                AASDiscoveryURL.value = url;
-                break;
-            case 'AASRegistry':
-                AASRegistryURL.value = url;
-                if (clearSelectedNode) useAASStore().dispatchSelectedNode({});
-                break;
-            case 'SubmodelRegistry':
-                SubmodelRegistryURL.value = url;
-                if (clearSelectedNode) useAASStore().dispatchSelectedNode({});
-                break;
-            case 'AASRepo':
-                AASRepoURL.value = url;
-                break;
-            case 'SubmodelRepo':
-                SubmodelRepoURL.value = url;
-                break;
-            case 'ConceptDescriptionRepo':
-                ConceptDescriptionRepoURL.value = url;
-                break;
-        }
-    }
-
     function dispatchDrawerState(dispatchedDrawerState: boolean): void {
         drawerState.value = dispatchedDrawerState;
     }
@@ -92,7 +126,7 @@ export const useNavigationStore = defineStore('navigationStore', () => {
         platform.value = dispatchedPlatform;
     }
 
-    function dispatchPlugins(dispatchedPlugins: Array<PluginType>): void {
+    function dispatchPlugins(dispatchedPlugins: PluginType[]): void {
         plugins.value = dispatchedPlugins;
     }
 
@@ -108,7 +142,150 @@ export const useNavigationStore = defineStore('navigationStore', () => {
         urlQuery.value = query;
     }
 
+    function dispatchComponentURL(componentKey: RepositoryKey, url: string, clearSelectedNode: boolean = true): void {
+        switch (componentKey) {
+            case 'AASDiscovery':
+                AASDiscoveryURL.value = url;
+                break;
+            case 'AASRegistry':
+                AASRegistryURL.value = url;
+                if (clearSelectedNode) aasStore.dispatchSelectedNode({});
+                break;
+            case 'SubmodelRegistry':
+                SubmodelRegistryURL.value = url;
+                if (clearSelectedNode) aasStore.dispatchSelectedNode({});
+                break;
+            case 'AASRepo':
+                AASRepoURL.value = url;
+                break;
+            case 'SubmodelRepo':
+                SubmodelRepoURL.value = url;
+                break;
+            case 'ConceptDescriptionRepo':
+                ConceptDescriptionRepoURL.value = url;
+                break;
+            default:
+                console.warn(`Unknown component key: ${componentKey}`);
+                break;
+        }
+    }
+
+    function connectComponents(): void {
+        // Ensure that Object.keys returns RepositoryKey[]
+        const keys = Object.keys(basyxComponents) as RepositoryKey[];
+
+        keys.forEach((repoKey) => {
+            const storedURL = window.localStorage.getItem(repoKey);
+
+            if (endpointConfigAvailable.value && storedURL) {
+                basyxComponents[repoKey].url = storedURL;
+                basyxComponents[repoKey].connect();
+            } else {
+                // Check environment path
+                let envPath = '';
+                switch (repoKey) {
+                    case 'AASDiscovery':
+                        envPath = AASDiscoveryURL.value;
+                        break;
+                    case 'AASRegistry':
+                        envPath = AASRegistryURL.value;
+                        break;
+                    case 'SubmodelRegistry':
+                        envPath = SubmodelRegistryURL.value;
+                        break;
+                    case 'AASRepo':
+                        envPath = AASRepoURL.value;
+                        break;
+                    case 'SubmodelRepo':
+                        envPath = SubmodelRepoURL.value;
+                        break;
+                    case 'ConceptDescriptionRepo':
+                        envPath = ConceptDescriptionRepoURL.value;
+                        break;
+                    default:
+                        break;
+                }
+
+                if (!basyxComponents[repoKey].url && envPath.trim() !== '') {
+                    basyxComponents[repoKey].url = envPath;
+                    basyxComponents[repoKey].connect();
+                }
+            }
+        });
+    }
+
+    async function connectComponent(repoKey: keyof typeof basyxComponents): Promise<void> {
+        const repo = basyxComponents[repoKey];
+        if (repo.url && repo.url.trim() !== '') {
+            repo.loading = true;
+            let path = repo.url;
+
+            // Append path check for backward compatibility if defined
+            if (repo.pathCheck && !path.includes(repo.pathCheck)) {
+                path += repo.pathCheck;
+            }
+
+            // Append additional parameters if any
+            if (repo.additionalParams) {
+                path += repo.additionalParams(repoKey);
+            }
+
+            const context = `connecting to ${repo.label}`;
+            const disableMessage = false;
+
+            try {
+                const response = await getRequest(path, context, disableMessage);
+                repo.loading = false;
+
+                if (response.success) {
+                    // Dispatch to the navigation store
+                    dispatchComponentURL(repoKey, repo.url);
+
+                    // Save to localStorage if endpoint config is available
+                    if (endpointConfigAvailable.value) {
+                        // console.log(`Saving ${repoKey} URL to localStorage:`, repo.url);
+                        window.localStorage.setItem(repoKey, repo.url);
+                    }
+
+                    // Update the connected status
+                    repo.connected = true;
+                } else {
+                    // Clear the URL in the navigation store
+                    dispatchComponentURL(repoKey, '');
+
+                    // Remove from localStorage if endpoint config is available
+                    if (endpointConfigAvailable.value) {
+                        // console.log(`Removing ${repoKey} URL from localStorage:`, repo.url);
+                        window.localStorage.removeItem(repoKey);
+                    }
+
+                    // Update the connected status
+                    repo.connected = false;
+                }
+            } catch (error) {
+                repo.loading = false;
+                console.error(`Error connecting to ${repo.label}:`, error);
+
+                // Clear the URL in the navigation store
+                dispatchComponentURL(repoKey, '');
+
+                // Remove from localStorage if endpoint config is available
+                if (endpointConfigAvailable.value) {
+                    // console.log(`Removing ${repoKey} URL from localStorage:`, repo.url);
+                    window.localStorage.removeItem(repoKey);
+                }
+
+                // Update the connected status
+                repo.connected = false;
+            }
+        } else {
+            repo.connected = false;
+            console.warn(`Repository URL for ${repoKey} is not defined or empty.`);
+        }
+    }
+
     return {
+        // Getters
         getDrawerState,
         getAASDiscoveryURL,
         getAASRegistryURL,
@@ -125,6 +302,9 @@ export const useNavigationStore = defineStore('navigationStore', () => {
         getTriggerAASListReload,
         getTriggerAASListScroll,
         getUrlQuery,
+        getBasyxComponents,
+
+        // Actions
         dispatchComponentURL,
         dispatchDrawerState,
         dispatchSnackbar,
@@ -136,5 +316,6 @@ export const useNavigationStore = defineStore('navigationStore', () => {
         dispatchTriggerAASListReload,
         dispatchTriggerAASListScroll,
         dispatchUrlQuery,
+        connectComponents,
     };
 });
